+++ src/components/mbx-heading/index.js
@@ -1,0 +1,3 @@
+import main from './mbx-heading';
+
+export default main;